incidence matrix renaming for dual/primal / E21 dual added in
Created E21 matrix for the dual grid. All operators are now present --> curl, divergence, grad. Simple verification that the curl of gradient = 0, divergence of curl = 0.

Added in a new class member that returns all the inner dual cells. Used for the dual E21 matrix. Documentation still needs to be added.